add security related stuff to project's composer
definition. The roave security-advisories conflicts with packages that
contain known security relevant issues and the sensiolabs security checker
does similar stuff with a call to an external API. The check is by default
done after composer install or update commands. Projects that don't want
to call an external API should remove the package from the require section
and remove the two calls in the post-install and post-update scripts.